+++ src/index.ts
@@ -3,9 +3,4 @@
  * See LICENSE in the project root for license information.
  */
 
-export * from "./crypto.js";
-export * from "./encoding.js";
-export * from "./model.js";
-export * from "./physical-binding.js";
-export * from "./replay.js";
-export * from "./session.js";
+export * from "./baseline/index.js";